Rearrange files to have a separate UI elements directory
