+++ src/UI/atoms/Icons/Star.tsx
