+++ src/components/Check/Check.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from '../_atoms/Button/Button';
+import Button from '../../UI/atoms/Button/Button';
 import '../Answers/Answers.css';
 import withAnimation from '../../HOCs/withAnimation/withAnimation';
